Move Controller to phasar-cli
