--- include/phasar/Controller.h
@@ -1,16 +1,0 @@
-/******************************************************************************
- * Copyright (c) 2023 Fabian Schiebel.
- * All rights reserved. This program and the accompanying materials are made
- * available under the terms of LICENSE.txt.
- *
- * Contributors:
- *     Fabian Schiebel and others
- *****************************************************************************/
-
-#ifndef PHASAR_CONTROLLER_H
-#define PHASAR_CONTROLLER_H
-
-#include "phasar/Controller/AnalysisController.h"
-#include "phasar/Controller/AnalysisControllerEmitterOptions.h"
-
-#endif // PHASAR_CONTROLLER_H
