+++ lib/CMakeLists.txt
@@ -4,7 +4,6 @@ add_subdirectory(PhasarPass)
 add_subdirectory(DB)
 add_subdirectory(Config)
 add_subdirectory(Utils)
-add_subdirectory(Controller)
 add_subdirectory(Pointer)
 add_subdirectory(ControlFlow)
 if(BUILD_PHASAR_CLANG)
@@ -40,7 +39,6 @@ add_phasar_library(phasar ${PHASAR_DYNLIB_KIND}
         phasar_llvm
         phasar_llvm_ifdside
         phasar_analysis_strategy
-        phasar_controller
     LINK_PRIVATE
         ${Boost_LIBRARIES}
     LLVM_LINK_COMPONENTS
