+++ tools/phasar-cli/Controller/AnalysisControllerInternalIDE.h
