+++ tools/phasar-cli/Controller/CMakeLists.txt
@@ -1,0 +1,3 @@
+file(GLOB_RECURSE CONTROLLER_SRC *.h *.cpp)
+
+target_sources(phasar-cli PRIVATE ${CONTROLLER_SRC})
